requirements.txt
bert_score==0.3.13
bottle==0.12.25
cchess==0.80
colorama==0.4.6
colorlog==6.7.0
cpm_kernels==1.0.11
editdistance==0.6.2
flair==0.12.2
Flask==2.3.3
gradio==3.44.3
h5py==3.8.0
jieba==0.42.1
keras==2.11.0
loralib==0.1.2
mdtex2html==1.2.0
nltk==3.8.1
numpy==1.21.6
onnx==1.14.1
onnxruntime==1.15.1
packaging==23.1
pandas==1.1.5
pynytimes==0.7.0
PyYAML==6.0
PyYAML==6.0.1
recommonmark==0.7.1
Requests==2.31.0
rouge==1.0.1
rouge_chinese==1.0.3
ruamel.base==1.0.0
sanic==23.6.0
scikit_learn==1.0.2
scipy==1.7.3
sentence_transformers==2.2.2
sentencepiece==0.1.98
setuptools==65.6.3
shap==0.41.0
six==1.16.0
sphinx_rtd_theme==1.3.0
sympy==1.12
tensorboardX==2.6.2.2
tensorflow==2.11.0
torch==1.13.1
torch4keras==0.0.6
torch_scatter==2.1.1
torchinfo==1.8.0
tqdm==4.65.0
transformers==4.28.1
tritonclient==2.37.0.9383150
